refactor: improve project archteture according to community doc
