+++ internal/pkg/validator/validator.go
@@ -1,7 +1,9 @@
-package common
+package validator
+
+import "luizalabs-technical-test/internal/pkg/formatter"
 
 // ValidateZipCode checks if the given string is a valid Brazilian ZipCode.
 func ValidateZipCode(zipCode string) bool {
-	zipCode = StripNonNumericCharacters(zipCode)
+	zipCode = formatter.StripNonNumericCharacters(zipCode)
 	return len(zipCode) == MinimumZipCodeLen
 }
