+++ internal/pkg/validator/validator_test.go
@@ -1,4 +1,4 @@
-package common
+package validator
 
 import (
 	"testing"
